Address clang-tidy fixes

+++ src/include/jank_analyze.h
@@ -5,7 +5,7 @@ struct jank_data
     int OOT = 0;
     int LOT = 0;
 
-    float nice() const;
+    double nice() const;
 };
 
 jank_data analyzeFrameData(const FtimeStamps &Fdata);
